% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fix_date_ymd}
\alias{fix_date_ymd}
\title{fix_date_ymd() — for converting YYYYMMDD date fields to YYYY-MM-DD}
\usage{
fix_date_ymd(date)
}
\arguments{
\item{date}{A character vector or numeric vector containing dates in \code{YYYYMMDD} format.}
}
\value{
A \code{Date} vector, with invalid values returned as \code{NA}.
}
\description{
Converts date strings in \code{YYYYMMDD} format to ISO \code{YYYY-MM-DD} format
and parses them into \code{Date} objects.
}
\examples{
fix_date_ymd("20250319")

}
