% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%not\%}
\alias{\%not\%}
\title{Frankie's custom not operator}
\usage{
... \%not\% NA
}
\arguments{
\item{x}{Vector of values to be tested}

\item{table}{Vector of values to be tested against}
}
\value{
Logical vector indicating whether each element of \code{x} is not in \code{table}
}
\description{
The \verb{\%not\%} operator returns the logical negation of the \code{\%in\%} operator.
It provides a more readable way to check if elements are not in a set.
}
\examples{
# Basic usage
c("apple", "banana", "cherry") \%not\% c("banana", "date")

# With dplyr
df <- data.frame(col = c("apple", "pear", "orange", "banana", "grape"))
df |> dplyr::filter(col \%not\% c("apple", "banana", "orange"))

}
