% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not.na}
\alias{not.na}
\title{Not NA}
\usage{
not.na(x, ...)
}
\arguments{
\item{x}{Vector of values to test}

\item{...}{Additional arguments (for consistency with other functions)}
}
\value{
Logical vector of the same length as \code{x}
}
\description{
Returns a logical vector indicating which elements are not NA.
This is the complement of \code{is.na()}.
}
\examples{
not.na(c(1, NA, 3, NA, 5))

}
